Apply code review suggestions
* Move tests into footnotes app
* Move factory class into separate file
* Remove unnecessary `response` fixture
