+++ .github/workflows/check-pr.yml
@@ -43,7 +43,7 @@ jobs:
     - name: Install Dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -r requirements-dev.txt
+        pip install -r requirements.txt
         npm install
 
     - name: Collect static assets
