+++ footnotes/views.py
@@ -3,7 +3,6 @@ from django.shortcuts import render
 from .models import Footnote
 
 
-# Create your views here.
 def list_footnotes(request):
     return render(
         request, "footnotes/list.jinja", context={"footnotes": Footnote.objects.all(),}
