+++ setup.cfg
@@ -25,7 +25,13 @@ force_single_line = True
 
 [flake8]
 exclude = .git,__pycache__,venv,*/migrations/*
-extend-ignore = E203
+extend-ignore =
+  D106, # don't require document strings for nested classes - usually Meta
+  E203, # black insists on [fn() : fn()] in slices - and pep8 recommends this
+  WPS306, # allow class without base class - we're not writing python2
+inline-quotes = " # black wants double quotes
+multiline-quotes = """
+docstring-quotes = """
 max-line-length = 88
 
 [tool:pytest]
